refactor(api): move auth logic to web api
It depends on the web framework (request).
